requirements.txt
Flask==3.0.0
Pillow==10.1.0
Flask-SQLAlchemy==3.1.1
python-slugify==8.0.1
markdown2==2.4.10
python-dotenv==1.0.0
Flask-WTF==1.2.1
Flask-Mail==0.9.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
